+++ src/frontend/apps/e2e/__tests__/app-impress/doc-editor.spec.ts
@@ -11,7 +11,12 @@ import {
   overrideConfig,
   verifyDocName,
 } from './utils-common';
-import { getEditor, openSuggestionMenu, writeInEditor } from './utils-editor';
+import {
+  getEditor,
+  mockAIResponse,
+  openSuggestionMenu,
+  writeInEditor,
+} from './utils-editor';
 import { connectOtherUserToDoc, updateShareLink } from './utils-share';
 import {
   createRootSubPage,
@@ -39,6 +44,7 @@ test.describe('Doc Editor', () => {
       .selectText();
 
     const toolbar = page.locator('.bn-formatting-toolbar');
+    await expect(toolbar.getByRole('button', { name: 'Ask AI' })).toBeVisible();
     await expect(
       toolbar.locator('button[data-test="comment-toolbar-button"]'),
     ).toBeVisible();
@@ -63,9 +69,6 @@ test.describe('Doc Editor', () => {
     ).toBeVisible();
     await expect(
       toolbar.locator('button[data-test="createLink"]'),
-    ).toBeVisible();
-    await expect(
-      toolbar.locator('button[data-test="ai-actions"]'),
     ).toBeVisible();
     await expect(
       toolbar.locator('button[data-test="convertMarkdown"]'),
@@ -93,13 +96,11 @@ test.describe('Doc Editor', () => {
 
     await expect(image).toHaveAttribute('role', 'presentation');
 
-    await image.dblclick();
+    await image.click();
 
+    await expect(toolbar.getByRole('button', { name: 'Ask AI' })).toBeHidden();
     await expect(
       toolbar.locator('button[data-test="comment-toolbar-button"]'),
-    ).toBeHidden();
-    await expect(
-      toolbar.locator('button[data-test="ai-actions"]'),
     ).toBeHidden();
     await expect(
       toolbar.locator('button[data-test="convertMarkdown"]'),
@@ -389,7 +390,117 @@ test.describe('Doc Editor', () => {
     await expect(image).toHaveAttribute('aria-hidden', 'true');
   });
 
-  test('it checks the AI buttons', async ({ page, browserName }) => {
+  test('it checks the AI feature and accepts changes', async ({
+    page,
+    browserName,
+  }) => {
+    await overrideConfig(page, {
+      AI_BOT: {
+        name: 'Albert AI',
+        color: '#8bc6ff',
+      },
+    });
+
+    await mockAIResponse(page);
+
+    await page.goto('/');
+
+    await createDoc(page, 'doc-ai', browserName, 1);
+
+    await openSuggestionMenu({ page });
+    await page.getByText('Ask AI').click();
+    await expect(
+      page.getByRole('option', { name: 'Continue Writing' }),
+    ).toBeVisible();
+    await expect(page.getByRole('option', { name: 'Summarize' })).toBeVisible();
+
+    await page.keyboard.press('Escape');
+
+    await page.locator('.bn-block-outer').last().fill('Hello World');
+
+    const editor = page.locator('.ProseMirror');
+    await editor.getByText('Hello World').selectText();
+
+    // Check from toolbar
+    await page.getByRole('button', { name: 'Ask AI' }).click();
+
+    await expect(
+      page.getByRole('option', { name: 'Improve Writing' }),
+    ).toBeVisible();
+    await expect(
+      page.getByRole('option', { name: 'Fix Spelling' }),
+    ).toBeVisible();
+    await expect(page.getByRole('option', { name: 'Translate' })).toBeVisible();
+
+    await page.getByRole('option', { name: 'Translate' }).click();
+    await page
+      .getByRole('textbox', { name: 'Ask anything...' })
+      .fill('Translate into french');
+    await page.getByRole('textbox', { name: 'Ask anything...' }).press('Enter');
+    await expect(editor.getByText('Albert AI')).toBeVisible();
+    await page
+      .locator('p.bn-mt-suggestion-menu-item-title')
+      .getByText('Accept')
+      .click();
+
+    await expect(editor.getByText('Bonjour le monde')).toBeVisible();
+
+    // Check Suggestion menu
+    await page.locator('.bn-block-outer').last().fill('/');
+    await expect(page.getByText('Write with AI')).toBeVisible();
+
+    // Reload the page to check that the AI change is still there
+    await page.goto(page.url());
+    await expect(editor.getByText('Bonjour le monde')).toBeVisible();
+  });
+
+  test('it reverts with the AI feature', async ({ page, browserName }) => {
+    await overrideConfig(page, {
+      AI_BOT: {
+        name: 'Albert AI',
+        color: '#8bc6ff',
+      },
+    });
+
+    await mockAIResponse(page);
+
+    await page.goto('/');
+
+    await createDoc(page, 'doc-ai', browserName, 1);
+
+    await page.locator('.bn-block-outer').last().fill('Hello World');
+
+    const editor = page.locator('.ProseMirror');
+    await editor.getByText('Hello World').selectText();
+
+    // Check from toolbar
+    await page.getByRole('button', { name: 'Ask AI' }).click();
+
+    await page.getByRole('option', { name: 'Translate' }).click();
+    await page
+      .getByRole('textbox', { name: 'Ask anything...' })
+      .fill('Translate into french');
+    await page.getByRole('textbox', { name: 'Ask anything...' }).press('Enter');
+    await expect(editor.getByText('Albert AI')).toBeVisible();
+    await expect(editor.getByText('Bonjour le monde')).toBeVisible();
+    await page
+      .locator('p.bn-mt-suggestion-menu-item-title')
+      .getByText('Revert')
+      .click();
+
+    await expect(editor.getByText('Hello World')).toBeVisible();
+  });
+
+  /**
+   * We have to skip this test for the CI for now, we cannot assert
+   * it because of `process.env.NEXT_PUBLIC_PUBLISH_AS_MIT` that is set
+   * at build time.
+   * It can be interesting to keep it for local tests.
+   */
+  test.skip('it checks the AI buttons feature', async ({
+    page,
+    browserName,
+  }) => {
     await page.route(/.*\/ai-translate\//, async (route) => {
       const request = route.request();
       if (request.method().includes('POST')) {
@@ -442,12 +553,77 @@ test.describe('Doc Editor', () => {
     await expect(editor.getByText('Bonjour le monde')).toBeVisible();
   });
 
+  /**
+   * We have to skip this test for the CI for now, we cannot assert
+   * it because of `process.env.NEXT_PUBLIC_PUBLISH_AS_MIT` that is set
+   * at build time.
+   * It can be interesting to keep it for local tests.
+   */
+  test.skip('it checks the AI buttons', async ({ page, browserName }) => {
+    await page.route(/.*\/ai-translate\//, async (route) => {
+      const request = route.request();
+      if (request.method().includes('POST')) {
+        await route.fulfill({
+          json: {
+            answer: 'Bonjour le monde',
+          },
+        });
+      } else {
+        await route.continue();
+      }
+    });
+
+    await createDoc(page, 'doc-ai', browserName, 1);
+
+    await page.locator('.bn-block-outer').last().fill('Hello World');
+
+    const editor = page.locator('.ProseMirror');
+    await editor.getByText('Hello').selectText();
+
+    await page.getByRole('button', { name: 'AI' }).click();
+
+    await expect(
+      page.getByRole('menuitem', { name: 'Use as prompt' }),
+    ).toBeVisible();
+    await expect(
+      page.getByRole('menuitem', { name: 'Rephrase' }),
+    ).toBeVisible();
+    await expect(
+      page.getByRole('menuitem', { name: 'Summarize' }),
+    ).toBeVisible();
+    await expect(page.getByRole('menuitem', { name: 'Correct' })).toBeVisible();
+    await expect(
+      page.getByRole('menuitem', { name: 'Language' }),
+    ).toBeVisible();
+
+    await page.getByRole('menuitem', { name: 'Language' }).hover();
+    await expect(
+      page.getByRole('menuitem', { name: 'English', exact: true }),
+    ).toBeVisible();
+    await expect(
+      page.getByRole('menuitem', { name: 'French', exact: true }),
+    ).toBeVisible();
+    await expect(
+      page.getByRole('menuitem', { name: 'German', exact: true }),
+    ).toBeVisible();
+
+    await page.getByRole('menuitem', { name: 'English', exact: true }).click();
+
+    await expect(editor.getByText('Bonjour le monde')).toBeVisible();
+  });
+
+  /**
+   * We have to skip this test for the CI for now, we cannot assert
+   * it because of `process.env.NEXT_PUBLIC_PUBLISH_AS_MIT` that is set
+   * at build time.
+   * It can be interesting to keep it for local tests.
+   */
   [
     { ai_transform: false, ai_translate: false },
     { ai_transform: true, ai_translate: false },
     { ai_transform: false, ai_translate: true },
   ].forEach(({ ai_transform, ai_translate }) => {
-    test(`it checks AI buttons when can transform is at "${ai_transform}" and can translate is at "${ai_translate}"`, async ({
+    test.skip(`it checks AI buttons when can transform is at "${ai_transform}" and can translate is at "${ai_translate}"`, async ({
       page,
       browserName,
     }) => {
@@ -520,6 +696,53 @@ test.describe('Doc Editor', () => {
         ).toBeHidden();
       }
     });
+  });
+
+  test(`it checks ai_proxy ability`, async ({ page, browserName }) => {
+    await mockedDocument(page, {
+      accesses: [
+        {
+          id: 'b0df4343-c8bd-4c20-9ff6-fbf94fc94egg',
+          role: 'owner',
+          user: {
+            email: 'super@owner.com',
+            full_name: 'Super Owner',
+          },
+        },
+      ],
+      abilities: {
+        destroy: true, // Means owner
+        link_configuration: true,
+        ai_proxy: false,
+        accesses_manage: true,
+        accesses_view: true,
+        update: true,
+        partial_update: true,
+        retrieve: true,
+      },
+      link_reach: 'restricted',
+      link_role: 'editor',
+      created_at: '2021-09-01T09:00:00Z',
+      title: '',
+    });
+
+    const [randomDoc] = await createDoc(
+      page,
+      'doc-editor-ai-proxy',
+      browserName,
+      1,
+    );
+
+    await verifyDocName(page, randomDoc);
+
+    await page.locator('.bn-block-outer').last().fill('Hello World');
+
+    const editor = page.locator('.ProseMirror');
+    await editor.getByText('Hello').selectText();
+
+    await expect(page.getByRole('button', { name: 'Ask AI' })).toBeHidden();
+    await page.locator('.bn-block-outer').last().fill('/');
+    await expect(page.getByText('Write with AI')).toBeHidden();
   });
 
   test('it downloads unsafe files', async ({ page, browserName }) => {