+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/BlockNoteSuggestionMenu.tsx
@@ -17,6 +17,7 @@ import {
   DocsStyleSchema,
 } from '../types';
 
+import BlockNoteAI from './AI';
 import {
   getCalloutReactSlashMenuItems,
   getPdfReactSlashMenuItems,
@@ -27,7 +28,13 @@ import XLMultiColumn from './xl-multi-column';
 const getMultiColumnSlashMenuItems =
   XLMultiColumn?.getMultiColumnSlashMenuItems;
 
-export const BlockNoteSuggestionMenu = () => {
+const getAISlashMenuItems = BlockNoteAI?.getAISlashMenuItems;
+
+export const BlockNoteSuggestionMenu = ({
+  aiAllowed,
+}: {
+  aiAllowed: boolean;
+}) => {
   const editor = useBlockNoteEditor<
     DocsBlockSchema,
     DocsInlineContentSchema,
@@ -50,6 +57,7 @@ export const BlockNoteSuggestionMenu = () => {
       getMultiColumnSlashMenuItems?.(editor) || [],
       getPdfReactSlashMenuItems(editor, t, fileBlocksName),
       getCalloutReactSlashMenuItems(editor, t, basicBlocksName),
+      aiAllowed && getAISlashMenuItems ? getAISlashMenuItems(editor) : [],
     );
 
     const index = combinedMenu.findIndex(
@@ -66,7 +74,14 @@ export const BlockNoteSuggestionMenu = () => {
 
     return async (query: string) =>
       Promise.resolve(filterSuggestionItems(newSlashMenuItems, query));
-  }, [basicBlocksName, editor, getInterlinkingMenuItems, t, fileBlocksName]);
+  }, [
+    editor,
+    t,
+    fileBlocksName,
+    basicBlocksName,
+    aiAllowed,
+    getInterlinkingMenuItems,
+  ]);
 
   return (
     <SuggestionMenuController